requirements.txt
streamlit
google-generativeai
python-dotenv
ipykernel
jupyter
chromadb
PyPDF2